requirements.txt
Pillow
numpy
protobuf
crccheck